+++ CleanCode/src/NuGet.Packager/Package.nuspec
@@ -2,7 +2,7 @@
 <package >
   <metadata>
     <id>CleanCode</id>
-    <version>4.0.0.0</version>
+    <version>4.0.0.2</version>
     <title>CleanCode</title>
     <authors>Hadi Hariri, Matt Ellis, SuperJMN</authors>
     <owners>Hadi Hariri, Matt Ellis</owners>
